Add User/Character-one/many, Character/Weapon-one/one, Character/Skill-many/many

+++ Services/CharacterService/CharacterService.cs
@@ -1,10 +1,12 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
-using _NET_Course.Dto;
+using _NET_Course.Dto.Character;
 using _NET_Course.Models;
 using AutoMapper;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace _NET_Course.Services
@@ -16,12 +18,16 @@ namespace _NET_Course.Services
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         /// <summary>
         /// Adds a new character to the database.
@@ -33,9 +39,13 @@ namespace _NET_Course.Services
             var _serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
             Character _character = _mapper.Map<Character>(character);
+            _character.User = await _context.Users.FirstOrDefaultAsync(u => u.ID == GetUserId());
+
             _context.Characters.Add(_character);
             await _context.SaveChangesAsync();
-            _serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            _serviceResponse.Data = await _context.Characters
+                .Where(c => c.ID == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
 
             return _serviceResponse;
         }
@@ -44,11 +54,11 @@ namespace _NET_Course.Services
         /// Gets all the characters from the database.
         /// </summary>
         /// <returns>The list of all characters in the database.</returns>
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userID)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var _serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
-            var _dbCharacters = await _context.Characters.Where(c => c.User.ID == userID).ToListAsync();
+            var _dbCharacters = await _context.Characters.Where(c => c.User.ID == GetUserId()).ToListAsync();
             _serviceResponse.Data = _dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
 
             return _serviceResponse;
@@ -63,7 +73,10 @@ namespace _NET_Course.Services
         {
             var _serviceResponse = new ServiceResponse<GetCharacterDto>();
 
-            var _dbCharacters = await _context.Characters.FirstOrDefaultAsync(c => c.ID == id);
+            var _dbCharacters = await _context.Characters
+                                .Include(c => c.Weapon)
+                                .Include(c => c.Skills)
+                                .FirstOrDefaultAsync(c => c.ID == id && c.User.ID == GetUserId());
             _serviceResponse.Data = _mapper.Map<GetCharacterDto>(_dbCharacters);
 
             return _serviceResponse;
@@ -80,18 +93,28 @@ namespace _NET_Course.Services
 
             try
             {
-                Character _character = await _context.Characters.FirstOrDefaultAsync(c => c.ID == character.ID);
+                Character _character = await _context.Characters
+                                        .Include(c => c.User)
+                                        .FirstOrDefaultAsync(c => c.ID == character.ID);
 
-                _character.Name = character.Name;
-                _character.HitPoints = character.HitPoints;
-                _character.Strength = character.Strength;
-                _character.Defence = character.Defence;
-                _character.Inteligence = character.Inteligence;
-                _character.Class = character.Class;
+                if(_character.User.ID == GetUserId())
+                {
+                    _character.Name = character.Name;
+                    _character.HitPoints = character.HitPoints;
+                    _character.Strength = character.Strength;
+                    _character.Defence = character.Defence;
+                    _character.Inteligence = character.Inteligence;
+                    _character.Class = character.Class;
 
-                await _context.SaveChangesAsync();
+                    await _context.SaveChangesAsync();
 
-                _serviceResponse.Data = _mapper.Map<GetCharacterDto>(_character);
+                    _serviceResponse.Data = _mapper.Map<GetCharacterDto>(_character);
+                }
+                else
+                {
+                    _serviceResponse.Success = false;
+                    _serviceResponse.Message = "Character not found";
+                }
             }
             catch (Exception e)
             {
@@ -113,11 +136,64 @@ namespace _NET_Course.Services
 
             try
             {
-                Character _character = _context.Characters.FirstOrDefault(c => c.ID == id);
-                _context.Characters.Remove(_character);
-                await _context.SaveChangesAsync();
+                Character _character = await _context.Characters.FirstOrDefaultAsync(c => c.ID == id && c.User.ID == GetUserId());
 
-                _serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                if(_character != null)
+                {
+                    _context.Characters.Remove(_character);
+                    await _context.SaveChangesAsync();
+
+                    _serviceResponse.Data = _context.Characters
+                                            .Where(c => c.User.ID == GetUserId())
+                                            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    _serviceResponse.Success = false;
+                    _serviceResponse.Message = "Character not found";
+                }
+            }
+            catch (Exception e)
+            {
+                _serviceResponse.Success = false;
+                _serviceResponse.Message = e.Message;
+            }
+
+            return _serviceResponse;
+        }
+
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto skill)
+        {
+            var _serviceResponse = new ServiceResponse<GetCharacterDto>();
+
+            try
+            {
+                var _character = await _context.Characters
+                                    .Include(c => c.Weapon)
+                                    .Include(c => c.Skills)
+                                    .FirstOrDefaultAsync(c => c.ID == skill.CharacterId && c.User.ID == GetUserId());
+                
+                if(_character == null)
+                {
+                    _serviceResponse.Success = false;
+                    _serviceResponse.Message = "Character not found.";
+
+                    return _serviceResponse;
+                }
+
+                var _skill = await _context.Skills.FirstOrDefaultAsync(s => s.ID == skill.SkillId);
+
+                if(_skill == null)
+                {
+                    _serviceResponse.Success = false;
+                    _serviceResponse.Message = "Skill not found.";
+
+                    return _serviceResponse;
+                }
+
+                _character.Skills.Add(_skill);
+                await _context.SaveChangesAsync();
+                _serviceResponse.Data = _mapper.Map<GetCharacterDto>(_character);
             }
             catch (Exception e)
             {
